tests/wpt/tests/fetch/metadata/generated/script-module-import-dynamic.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/script-module-import-dynamic.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for dynamic ECMAScript module import</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <script type="module">
  'use strict';

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - Not sent to non-trustworthy cross-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-mode');
        });
  }, 'sec-fetch-mode - Not sent to non-trustworthy cross-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-dest');
        });
  }, 'sec-fetch-dest - Not sent to non-trustworthy cross-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-user');
        });
  }, 'sec-fetch-user - Not sent to non-trustworthy cross-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy same-origin destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpSameSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy same-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpCrossSite'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-storage-access');
        });
  }, 'sec-fetch-storage-access - Not sent to non-trustworthy cross-site destination');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsOrigin', 'httpOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade (header not sent)');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpOrigin', 'httpsOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - HTTPS upgrade');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['httpsOrigin', 'httpOrigin', 'httpsOrigin'], { mime: 'application/javascript' }
    );

    return import(url)
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_array_equals(headers['sec-fetch-site'], ['cross-site']);
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
</html>
